커뮤 댓글 작업중

+++ commu/views.py
@@ -1,12 +1,8 @@
 from django.contrib.auth.decorators import login_required
-from django.http import JsonResponse
 from django.shortcuts import render, redirect
 from commu.models import Article
 from commu.models import Comment
-from commu.services.article_service import create_an_article, get_article_list
-from commu.services.comment_service import create_an_comment, delete_an_comment, update_an_comment
-
-
+from commu.services.comment_service import create_an_comment, delete_an_comment, update_an_comment, get_comment_page
 
 
 # @login_required(login_url:'sign_in')
@@ -18,31 +14,50 @@ def commu_view(request):
 
 # @login_required(login_url:'sign_in')
 def write_comment(request, id): #해당게시글의 id
-    if request.method == 'POST': #댓글작성 요청
-        user = request.user # 지금 접속중인 user정보
-        comment = request.POST['comment']
-        article = Article.objects.get(id=id)
-        comment = create_an_comment(article, user, comment)
-        return redirect() #해당게시글 페이지로
+    user = request.user  # 지금 접속중인 user정보
+    comment = request.POST['comment']
+    article = Article.objects.get(id=id)
 
-      
+    if request.method == 'POST': #댓글작성 요청
+        if comment != "":#댓글 빈칸아니면
+            create_an_comment(article, user, comment)
+            return redirect('/commu/' + str(id)) #해당게시글 페이지로
+        else: #빈칸이면
+            return render(request,'commu/commu.html', {'article': article} )
+
 # @login_required(login_url:'sign_in')
 def update_comment(request, id): #해당댓글의 id
+    comment = Comment.objects.filter(id=id).get()  # 쿼리셋
+    article_id = comment.article_id
     if request.method == 'POST':
-        comment = Comment.objects.filter(id=id).get() #쿼리셋
         edit_comment = request.POST['update_comment'] # 댓글모델에 바꿀 댓글내용
+        print(edit_comment)
         if request.user == comment.author: #접속중인 유저가 댓글작성자라면
-            # 댓글id와, 바꿀 해당쿼리셋객체, 수정할 댓글내용을 넘겨줌
-            update_an_comment(comment, edit_comment) #댓글 수정함수 #메세지 넘길거면 return값 넘기기
-            return redirect()#해당게시글 페이지로
-        else:
-            return JsonResponse({'msg':'잘못된 접근입니다.'})#어떤 방식으로 프론트에서 보여주냐에 따라 바뀔수 있음
+            if edit_comment != "":
+                # 댓글id와, 바꿀 해당쿼리셋객체, 수정할 댓글내용을 넘겨줌
+                update_an_comment(comment, edit_comment) #댓글 수정함수 #메세지 넘길거면 return값 넘기기
+                print('성공')
+                return redirect('/commu/' + str(article_id))#해당게시글 페이지로
+            else:#빈칸이면
+                print('빈칸')
+                from django.contrib import messages
+                message = messages.info(request, '내용을 입력하세요.')
+                return redirect('/commu/' + str(article_id),messages=message)#해당게시글 페이지로
+        else:#메세지 띄우기
+            return redirect('/commu/' + str(article_id))#해당게시글 페이지로
+    else:
+        return redirect('/commu/' + str(article_id))
 
           
 # @login_required(login_url:'sign_in')
 def delete_comment(request, id):# 해당댓글 id
-    delete_an_comment(id)
-    return redirect()#해당 게시글 페이지로
+    comment = Comment.objects.filter(id=id).get()
+    article_id = comment.article_id
+    if request.user == comment.author:
+        delete_an_comment(id)
+        return redirect('/commu/' + str(article_id))#해당 게시글 페이지로
+    else:
+        return redirect('commu/' + str(article_id))
 
 
 def article_create(request):
@@ -62,7 +77,9 @@ def article_create(request):
 
 def article_detail(request, id):
     get_article = Article.objects.get(id=id)
-    return render(request, 'commu/commu_detail.html', {'article': get_article})
+    page = request.GET.get("page")
+    comments = get_comment_page(page, id)
+    return render(request, 'commu/commu_detail.html', {'article': get_article, 'comments': comments})
 
 
 def delete_an_article(request, id): # 글 삭제
